webserver: request handler fix bug
Each and every request that has len have to be read entirely.
If not it has the same data pointer for future read. That cause when new
packet receive we have to know that current read is read of previous
packet. So always read entire packet.

Fix ping response. And refactor handler base on Dmitry's code.

Signed-off-by: vmitrofanov <42731144+vmitrofanov@users.noreply.github.com>